+++ apps/api-service/src/auth/auth.service.ts
@@ -11,7 +11,7 @@ import {
 import { prodUIURI } from "@/shared/constants/other-constants"
 import { statusMessages } from "@/shared/constants/status-messages"
 import { EventEmitter2, OnEvent } from "@nestjs/event-emitter"
-import { EventMap } from "@/shared/utils/event.map"
+import { EventMap } from "@/shared/constants/event.map"
 import { CommandBus, QueryBus } from "@nestjs/cqrs"
 import { FindUserByEmailQuery } from "./queries/impl/find-user-by-email.query"
 import { User } from "./schemas/user.schema"
@@ -178,10 +178,7 @@ export class AuthService {
           subscription = subscriptionRes.shift()
         }
 
-        const isSubscriptionActive =
-          subscription && new Date(subscription.endsAt) > new Date()
-
-        return { user, subscription, isSubscriptionActive }
+        return { user, subscription }
       } else {
         throw new Error(statusMessages.invalidUser)
       }