+++ apps/api-service/src/auth/repositories/user.repository.ts
@@ -1,10 +1,10 @@
 import { Injectable } from "@nestjs/common"
 import { InjectModel } from "@nestjs/mongoose"
 import { User } from "../schemas/user.schema"
-import { GeneralDbConnectionMap } from "src/shared/utils/db-connection.map"
+import { GeneralDbConnectionMap } from "@/shared/constants/db-connection.map"
 import { FilterQuery, Model } from "mongoose"
 import { OnEvent } from "@nestjs/event-emitter"
-import { EventMap } from "@/shared/utils/event.map"
+import { EventMap } from "@/shared/constants/event.map"
 import { EntityRepository } from "@/shared/entity/entity.repository"
 import objectId from "@/shared/utils/convert-objectid"