+++ apps/api-service/src/platform/subscription/subscription.service.ts
@@ -3,12 +3,12 @@ import { Injectable } from "@nestjs/common"
 import { statusMessages } from "@/shared/constants/status-messages"
 import { config } from "src/config"
 import { OnEvent } from "@nestjs/event-emitter"
-import { EventMap } from "@/shared/utils/event.map"
+import { EventMap } from "@/shared/constants/event.map"
 import { CommandBus, QueryBus } from "@nestjs/cqrs"
 import { CreateSubscriptionCommand } from "./commands/impl/create-subscription.command"
 import { FindSubscriptionByUserIdQuery } from "./queries/impl/find-subscription-by-user-id.query"
 import { getRediretURIAPI } from "./utils/redirect-uri"
-import { subscriptionConfig } from "../config/data/subscription.config"
+import { Subscription } from "./schemas/subscription.schema"
 
 @Injectable()
 export class SubscriptionService {
@@ -25,7 +25,7 @@ export class SubscriptionService {
     userId: string
   ): Promise<Stripe.Checkout.Session> {
     try {
-      const { price } = subscriptionConfig
+      const price = 50
       const session = await this.stripe.checkout.sessions.create({
         payment_method_types: ["card"],
         line_items: [
@@ -82,9 +82,24 @@ export class SubscriptionService {
   @OnEvent(EventMap.GetSubscriptionDetails)
   async getMySubscription(userId: string) {
     try {
-      return await this.queryBus.execute(
-        new FindSubscriptionByUserIdQuery(userId)
-      )
+      const subscription: Subscription | null | undefined =
+        await this.queryBus.execute(new FindSubscriptionByUserIdQuery(userId))
+
+      if (!subscription) {
+        return null
+      }
+
+      const { _id, price, endsAt } = subscription
+      const isActive = subscription && new Date(endsAt) > new Date()
+
+      return {
+        _id,
+        userId,
+        price,
+        createdAt: (subscription as any).createdAt,
+        endsAt,
+        isActive,
+      }
     } catch (error) {
       throw new Error(statusMessages.connectionError)
     }