+++ apps/api-service/src/products/intelligence/intelligence.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from "@nestjs/common"
 import { InjectModel } from "@nestjs/mongoose"
-import { ProductsDbConnectionMap } from "src/shared/utils/db-connection.map"
+import { ProductsDbConnectionMap } from "@/shared/constants/db-connection.map"
 import { Thread } from "./schemas/thread.schema"
 import { Model } from "mongoose"
 import { EntityRepository } from "@/shared/entity/entity.repository"