+++ apps/api-service/src/products/taxadvisor/taxadvisor.strategy.ts
@@ -2,11 +2,12 @@ import { Injectable } from "@nestjs/common"
 import { Thread } from "./schemas/thread.schema"
 import { config } from "@/config"
 import { ChatOpenAI } from "@langchain/openai"
+import { PromptTemplate } from "@langchain/core/prompts"
 import { ChatGoogleGenerativeAI } from "@langchain/google-genai"
 import { createReactAgent } from "@langchain/langgraph/prebuilt"
 import { LanguageModelLike } from "@langchain/core/language_models/base"
 import { User } from "@/auth/schemas/user.schema"
-import { taxAdvisorSystemPrompt } from "./data/tax-advisor-system-prompt"
+import { RedisService } from "@/shared/redis/redis.service"
 
 export interface TaxAdvisorStrategyType {
   genericName: string
@@ -20,13 +21,25 @@ export interface TaxAdvisorStrategyType {
 
 @Injectable()
 export class TaxAdvisorStrategy {
-  constructor() {}
+  constructor(private readonly redisService: RedisService) {}
+
+  private async getSystemInstruction(user: User) {
+    const data = await this.redisService.get("taxadvisor-system-instruction")
+    return PromptTemplate.fromTemplate(data).invoke({
+      appName: config.APP_NAME,
+      userName: user.name,
+      userId: user.id,
+      userEmail: user.email,
+      baseCurrency: user.baseCurrency,
+    })
+  }
 
   private async runAdvisorAgent(
     llm: LanguageModelLike,
     args: TaxAdvisorStrategyType
   ) {
     const { thread, prompt, user } = args
+    const systemInstruction = await this.getSystemInstruction(user)
 
     const agent = createReactAgent({
       llm,
@@ -40,7 +53,7 @@ export class TaxAdvisorStrategy {
 
     const { messages } = await agent.invoke({
       messages: [
-        { role: "system", content: taxAdvisorSystemPrompt(user) },
+        { role: "system", content: systemInstruction.value },
         ...chatHistory,
         { role: "user", content: prompt },
       ],