+++ apps/api-service/src/products/wealthanalyzer/institution/institution.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@nestjs/common"
 import { InjectModel } from "@nestjs/mongoose"
 import { Institution } from "./schemas/institution.schema"
-import { ProductsDbConnectionMap } from "src/shared/utils/db-connection.map"
+import { ProductsDbConnectionMap } from "@/shared/constants/db-connection.map"
 import { Model } from "mongoose"
 import { EntityRepository } from "@/shared/entity/entity.repository"