+++ apps/ui-service/src/app/(platform)/dashboard/page.tsx
@@ -5,11 +5,19 @@ import LiabilityCard from "@/shared/components/dashboard-cards/liability-card"
 import WealthCard from "@/shared/components/dashboard-cards/wealth-card"
 import { ProductCard } from "@/shared/components/marketing-cards"
 import { endPoints } from "@/shared/constants/api-endpoints"
+import { uiConstants } from "@/shared/constants/global-constants"
 import HTTPMethods from "@/shared/constants/http-methods"
+import notify from "@/shared/hooks/use-notify"
 import useQuery from "@/shared/hooks/use-query"
 import { ProductsConfig } from "@/shared/types"
+import { useSearchParams } from "next/navigation"
+import { useRouter } from "nextjs-toploader/app"
+import { useEffect } from "react"
 
 export default function Page() {
+  const searchParams = useSearchParams()
+  const router = useRouter()
+
   const { data } = useQuery<ProductsConfig>({
     queryKey: ["getProductConfig"],
     queryUrl: endPoints.getProductConfig,
@@ -21,6 +29,20 @@ export default function Page() {
       <ProductCard key={product.productName} product={product} />
     ))
   }
+
+  useEffect(() => {
+    const subscriptionSuccess = searchParams.get("subscriptionSuccess")
+    if (subscriptionSuccess !== null) {
+      if (subscriptionSuccess === "true") {
+        notify(uiConstants.subscriptionSuccess, "success")
+      }
+
+      if (subscriptionSuccess === "false") {
+        notify(uiConstants.subscriptionFailed, "error")
+      }
+      router.push("/dashboard")
+    }
+  }, [searchParams])
 
   return (
     <div className="mx-auto grid w-full items-start gap-6">