+++ apps/ui-service/src/app/(platform)/products/wealthanalyzer/dashboard/page.tsx
@@ -7,19 +7,12 @@ import {
   InstitutionCard,
   AddInstitutionCard,
 } from "@/shared/components/institutioncard"
-import { useRouter } from "nextjs-toploader/app"
-import { useEffect } from "react"
-import notify from "@/shared/hooks/use-notify"
-import { uiConstants } from "@/shared/constants/global-constants"
-import { useSearchParams } from "next/navigation"
 import LiabilityCard from "@/shared/components/dashboard-cards/liability-card"
 import WealthCard from "@/shared/components/dashboard-cards/wealth-card"
 import GoalDashboardCard from "@/shared/components/dashboard-cards/goal-dashboard-card"
 import EMICard from "@/shared/components/dashboard-cards/emi-card"
 
 export default function Page() {
-  const searchParams = useSearchParams()
-  const router = useRouter()
   const institutions = useQuery<Institution[]>({
     queryKey: ["get-institutions"],
     queryUrl: endPoints.institution,
@@ -29,20 +22,6 @@ export default function Page() {
   const renderInstitutions = institutions?.data?.map((institution) => (
     <InstitutionCard institution={institution} key={institution._id} />
   ))
-
-  useEffect(() => {
-    const subscriptionSuccess = searchParams.get("subscriptionSuccess")
-    if (subscriptionSuccess !== null) {
-      if (subscriptionSuccess === "true") {
-        notify(uiConstants.subscriptionSuccess, "success")
-      }
-
-      if (subscriptionSuccess === "false") {
-        notify(uiConstants.subscriptionFailed, "error")
-      }
-      router.push("/dashboard")
-    }
-  }, [searchParams])
 
   return (
     <div className="mx-auto grid w-full items-start gap-6">