+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.4
+
+- Moved the configurations to redis
+- Redis integration for config
+- Standardized error handling
+
 ## v1.3
 
 - Minor changes & bug fixes on attribute updates